fix(updater): close release cut transition authority
